Updated testing_directory function to allow for parametrisation
